[FLOW-77] Double loading bug fix and Refactoring.
