+++ app/views/dashboard/project_flow_health.json.jbuilder
@@ -1,0 +1,4 @@
+json.remaining_days @vpi[:remaining_days]
+json.remaining_issues @vpi[:remaining_issues]
+json.average_time_to_close @vpi[:average_time_to_close]
+json.collapsable current_user.vpi
