+++ app/views/dashboard/projects.json.jbuilder
@@ -1,0 +1,6 @@
+json.projects @projects do |project|
+  json.project_id project.project_id
+  json.key project.key
+  json.name project.name
+end
+json.sortable_items current_user.sortable_items
