+++ app/views/dashboard/projects_vpi.json.jbuilder
@@ -1,0 +1,1 @@
+json.vpi_by_project @vpi_by_project
