+++ app/views/dashboard/value_stream_map.json.jbuilder
@@ -1,0 +1,5 @@
+json.lead_time @lead_time
+json.process_time @process_time
+json.percent_c_a @percent_c_a
+json.total @total
+json.collapsable current_user.vsm
