+++ app/views/users/sessions/pre_otp.json.jbuilder
@@ -3,7 +3,8 @@ unless @user.two_factor_auth.present?
   json.user do
     json.extract! @user, :id, :email, :two_factor_auth, :first_name,
       :last_name, :username, :company_name, :jira_url, :api_version,
-      :jira_username, :jira_password, :initial_config
+      :jira_username, :jira_password, :initial_config, :wip, :gauge, :focus, :vpi,
+      :activities, :vsm, :sortable_items
     json.avatar url_for(@user.avatar) if @user.avatar.present?
   end
 end
