+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_04_15_123537) do
+ActiveRecord::Schema.define(version: 2021_05_01_093631) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -109,6 +109,12 @@ ActiveRecord::Schema.define(version: 2021_04_15_123537) do
     t.json "min_max"
     t.json "sortable_items"
     t.boolean "initial_config"
+    t.boolean "wip"
+    t.boolean "gauge"
+    t.boolean "focus"
+    t.boolean "vpi"
+    t.boolean "activities"
+    t.boolean "vsm"
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
