+++ front/js/components/Dashboard/Dashboard.js
@@ -1,915 +1,84 @@
 import React, { Component } from 'react';
-import { NavLink } from 'react-router-dom';
-import { Doughnut, Bar, Bubble } from 'react-chartjs-2';
-import * as ChartAnnotation from 'chartjs-plugin-annotation';
-import GaugeChart from 'react-gauge-chart';
-import {
-  Container,
-  Row,
-  Col,
-  Card,
-  ListGroup,
-  Badge,
-  Image,
-  OverlayTrigger,
-  Tooltip,
-  ProgressBar,
-  Form,
-  Alert,
-  Spinner,
-  Modal,
-  Collapse,
-  Dropdown,
-  Figure,
-  Table,
-  Button,
-} from 'react-bootstrap';
-import Skeleton from 'react-loading-skeleton';
-import InfiniteScroll from 'react-infinite-scroll-component';
-import axios from 'axios';
 import { Redirect } from 'react-router';
-import moment from 'moment';
-import {
-  SortableContainer,
-  SortableElement,
-  SortableHandle,
-} from 'react-sortable-hoc';
+import { Container, Row, Col, Form, Alert } from 'react-bootstrap';
+import axios from 'axios';
+import { SortableContainer, SortableElement } from 'react-sortable-hoc';
 import arrayMove from 'array-move';
-import formattedTime from '../../utils/timeFormatter';
 import Footer from '../Shared/Footer';
+import Activities from './Activities';
+import VSM from './VSM';
+import VPI from './VPI';
+import Focus from './Focus';
+import Gauge from './Gauge';
+import WIP from './WIP';
+import Navbar from './Navbar';
 
 export default class Dashboard extends Component {
   constructor(props) {
     super(props);
+    const defaultItems = ['vpi', 'wip', 'gauge', 'focus', 'activities', 'vsm'];
+    const user = JSON.parse(localStorage.getItem('user'));
+    const sortableItems = user.sortable_items;
+    const initialConfig = user.initial_config;
     this.state = {
-      issues: [],
-      selectedIssueId: '',
-      epics: [],
-      epicIssues: [],
-      bugs: [],
-      tasks: [],
       projects: [],
-      devArcLength: [0.7, 0.3],
-      testArcLength: [0.2, 0.8],
-      deployArcLength: [0.15, 0.85],
-      redirect: false,
-      project_id: '',
-      jiraActivityLoading: false,
-      jiraActivityHasMore: true,
+      projectId: '',
       showAlert: false,
       alertMessage: '',
-      projectsSync: false,
-      issuesSync: false,
-      leadTime: '',
-      hoveredFocus: '',
-      remaining_days: null,
-      remaining_issues: 0,
-      average_time_to_close: null,
-      showVPI: false,
-      testVPI: 1,
-      testAverageTimeToClose: 1,
-      testRemainingIssues: 1,
-      testRemainingDays: 1,
-      totalBacklog: 0,
-      totalWorkInProgress: 0,
-      totalWorkInReview: 0,
-      backlogLimit: 0,
-      inProgressLimit: 0,
-      inReviewLimit: 0,
-      WIPChartType: 'bar',
-      showWIPSection: true,
-      showGaugeSection: true,
-      showFocusSection: true,
-      showVPISection: true,
-      showActivitiesSection: true,
-      showVSMSection: true,
-      sortableItems: ['vpi', 'wip', 'gauge', 'focus', 'activities', 'vsm'],
-      user: {},
-      showProjectVPIs: false,
-      showIngest: false,
-      projectVPIs: [],
-      statusLeadTimes: {},
-      statusProcessTimes: {},
-      statusCA: {},
-      VSMTotal: {},
-      ingest: null,
-      ingestType: 'projects',
-      showIngestAlert: false,
-      ingestErrorMessage: '',
-      disableIngest: false,
+      sortableItems: sortableItems === null ? defaultItems : sortableItems,
+      initialConfig: initialConfig === true,
     };
   }
 
   componentDidMount = () => {
     this.fetchInitData();
-    const user = JSON.parse(localStorage.getItem('user'));
-    this.setState({ user });
   };
 
-  fetchInitData = (project_id = '') => {
-    this.setState({ jiraActivityLoading: true, jiraActivityHasMore: true });
-
+  fetchInitData = () => {
     axios
-      .get(`/jiras.json?start_at=0&project_id=${project_id}`)
+      .get(`/dashboard/projects.json`)
       .then((response) => {
-        this.setState({ issues: response.data, jiraActivityLoading: false });
-      })
-      .catch((error) => {
-        const { message } = error.response.data;
+        const { data } = response;
+        const { sortableItems } = this.state;
+
         this.setState({
-          jiraActivityLoading: false,
-          showAlert: true,
-          alertMessage: message,
+          projects: data.projects,
+          sortableItems:
+            data.sortable_items === null ? sortableItems : data.sortable_items,
         });
       })
-      .finally(() => {
-        const { issues } = this.state;
-        if (issues.length === 0) {
-          this.setState({ jiraActivityHasMore: false });
-        } else {
-          this.setState({ selectedIssueId: issues[0].id });
-        }
-        setTimeout(() => {
-          this.setState({ showAlert: false });
-        }, 3000);
-      });
-
-    axios.get(`/jiras/stat.json?project_id=${project_id}`).then((response) => {
-      const { data } = response;
-      const {
-        total_backlog,
-        total_in_progress,
-        total_done,
-        grand_total,
-        epics,
-        epic_issues,
-        bugs,
-        tasks,
-        projects,
-        remaining_days,
-        remaining_issues,
-        average_time_to_close,
-        total_work_in_progress,
-        total_work_in_review,
-        boards,
-        min_max,
-        sortable_items,
-        vpi_by_project,
-        lead_time,
-        process_time,
-        percent_c_a,
-        total,
-      } = data;
-      let devPercent;
-      let devPendingPercent;
-      let testPercent;
-      let testPendingPercent;
-      let deployPercent;
-      let deployPendingPercent;
-      let backlogLimit = 0;
-      let inProgressLimit = 0;
-      let inReviewLimit = 0;
-      let showWIPSection = true;
-      let showGaugeSection = true;
-      let showFocusSection = true;
-      let showVPISection = true;
-      let showActivitiesSection = true;
-      let showVSMSection = true;
-      let sortableItems = ['wip', 'gauge', 'focus', 'vpi', 'activities'];
-      const leadStatus = ['backlog', 'to do', 'open'];
-      const devStatus = ['selected for development', 'in progress'];
-      const revStatus = ['review', 'qa', 'ready for review', 'in review'];
-
-      if (grand_total > 0) {
-        devPercent = total_backlog / grand_total;
-        devPendingPercent = (grand_total - total_backlog) / grand_total;
-        testPercent = total_in_progress / grand_total;
-        testPendingPercent = (grand_total - total_in_progress) / grand_total;
-        deployPercent = total_done / grand_total;
-        deployPendingPercent = (grand_total - total_done) / grand_total;
-      }
-
-      if (boards.length > 0) {
-        boards[0].column_config.forEach((config) => {
-          if (config.statuses.length > 0 && config.max !== undefined) {
-            if (leadStatus.includes(config.name.toLowerCase())) {
-              backlogLimit += config.max;
-            } else if (devStatus.includes(config.name.toLowerCase())) {
-              inProgressLimit += config.max;
-            } else if (revStatus.includes(config.name.toLowerCase())) {
-              inReviewLimit += config.max;
-            }
-          }
-        });
-      }
-
-      if (min_max !== null) {
-        showWIPSection = min_max.showWIPSection;
-        showGaugeSection = min_max.showGaugeSection;
-        showFocusSection = min_max.showFocusSection;
-        showVPISection = min_max.showVPISection;
-        showActivitiesSection = min_max.showActivitiesSection;
-        showVSMSection = min_max.showVSMSection;
-      }
-
-      if (sortable_items !== null) {
-        sortableItems = sortable_items;
-      }
-
-      this.setState({
-        devArcLength: [devPercent, devPendingPercent],
-        testArcLength: [testPercent, testPendingPercent],
-        deployArcLength: [deployPercent, deployPendingPercent],
-        epics,
-        epicIssues: epic_issues,
-        bugs,
-        tasks,
-        projects,
-        remaining_days,
-        remaining_issues,
-        average_time_to_close,
-        totalBacklog: total_backlog,
-        totalWorkInProgress: total_work_in_progress,
-        totalWorkInReview: total_work_in_review,
-        backlogLimit,
-        inProgressLimit,
-        inReviewLimit,
-        showWIPSection,
-        showGaugeSection,
-        showFocusSection,
-        showVPISection,
-        showActivitiesSection,
-        showVSMSection,
-        sortableItems,
-        projectVPIs: vpi_by_project,
-        statusLeadTimes: lead_time,
-        statusProcessTimes: process_time,
-        statusCA: percent_c_a,
-        VSMTotal: total,
-      });
-    });
-  };
-
-  fetchMoreData = () => {
-    const { issues, project_id } = this.state;
-    const startAt = issues.length;
-
-    axios
-      .get(`/jiras.json?start_at=${startAt}&project_id=${project_id}`)
-      .then((response) => {
-        this.setState({ issues: issues.concat(response.data) });
-        if (response.data.length === 0) {
-          this.setState({ jiraActivityHasMore: false });
-        }
-      })
-      .catch((error) => {
-        const { message } = error.response.data;
-        this.setState({ showAlert: true, alertMessage: message });
-      })
-      .finally(() => {
-        setTimeout(() => {
-          this.setState({ showAlert: false });
-        }, 3000);
-      });
-  };
-
-  logoutUser = () => {
-    axios.delete('/users/sign_out.json').then(() => {
-      localStorage.removeItem('auth_token');
-      localStorage.removeItem('user');
-      this.setState({ redirect: true });
-    });
+      .catch(() => {});
   };
 
   handleProjectChange = (event) => {
     const { value } = event.target;
-    this.setState({ project_id: value });
-    this.fetchInitData(value);
-  };
-
-  handleIngestTypeChange = (event) => {
-    this.setState({ ingestType: event.target.value });
-  };
-
-  handleIngestChange = (event) => {
-    this.setState({ ingest: event.target.files[0] });
-  };
-
-  handleSubmit = (e) => {
-    e.preventDefault();
-    const { ingest, ingestType } = this.state;
-    const data = new FormData();
-    data.append(`user[${ingestType}_ingest]`, ingest);
-    this.setState({ disableIngest: true });
-
-    axios
-      .put('/users/ingest.json', data)
-      .then((response) => {
-        window.location.href = '/';
-      })
-      .catch((error) => {
-        const { message } = error.response.data;
-        this.setState({ showIngestAlert: true, ingestErrorMessage: message });
-      })
-      .finally(() => {
-        this.setState({ disableIngest: false });
-        setTimeout(() => {
-          this.setState({ showIngestAlert: false });
-        }, 3000);
-      });
-  };
-
-  selectIssue = (issueId) => {
-    this.setState({ selectedIssueId: issueId });
-  };
-
-  clearLeadTime = (event) => {
-    event.stopPropagation();
-    this.setState({ hoveredFocus: '', leadTime: '' });
-  };
-
-  handleVPIClose = () => this.setState({ showVPI: false });
-
-  handleVPIShow = () => this.setState({ showVPI: true });
-
-  handleProjectVPIsClose = () => this.setState({ showProjectVPIs: false });
-
-  handleProjectVPIsShow = () => this.setState({ showProjectVPIs: true });
-
-  handleIngestClose = () =>
-    this.setState({ showIngest: false, ingestType: 'projects', ingest: null });
-
-  handleIngestShow = () => this.setState({ showIngest: true });
-
-  handleRD = (event) => {
-    const { value } = event.target;
-    if (value === '') return;
-    this.setState({ testRemainingDays: value });
-    this.updateVPI('rd', value);
-  };
-
-  handleRI = (event) => {
-    const { value } = event.target;
-    if (value === '') return;
-    this.setState({ testRemainingIssues: value });
-    this.updateVPI('ri', value);
-  };
-
-  handleACR = (event) => {
-    const { value } = event.target;
-    if (value === '') return;
-    this.setState({ testAverageTimeToClose: value });
-    this.updateVPI('acr', value);
-  };
-
-  updateVPI = (param, value) => {
-    let {
-      testAverageTimeToClose,
-      testRemainingIssues,
-      testRemainingDays,
-    } = this.state;
-
-    if (param === 'rd') {
-      testRemainingDays = value;
-    } else if (param === 'ri') {
-      testRemainingIssues = value;
-    } else if (param === 'acr') {
-      testAverageTimeToClose = value;
-    }
-
-    const testVPI = (
-      testRemainingDays /
-      (testAverageTimeToClose * testRemainingIssues)
-    ).toFixed(2);
-
-    this.setState({ testVPI });
-  };
-
-  WIPChartTypeChangeHandler = (event) => {
-    this.setState({ WIPChartType: event.target.id });
-  };
-
-  syncProjects = () => {
-    this.setState({ projectsSync: true });
-
-    axios
-      .put('/users/sync_projects.json', {})
-      .then((response) => {
-        const { projects } = response.data;
-        this.setState({ projects });
-      })
-      .finally(() => {
-        this.setState({ projectsSync: false });
-      });
-  };
-
-  syncIssues = () => {
-    this.setState({ issuesSync: true });
-
-    axios
-      .put('/users/sync_issues.json', {})
-      .then((response) => {
-        window.location.href = '/';
-      })
-      .finally(() => {
-        this.setState({ issuesSync: false });
-      });
-  };
-
-  displayLeadTime = (epicId) => {
-    this.setState({ hoveredFocus: epicId, leadTime: 'Loading...' });
-
-    axios
-      .get(`/jiras/${epicId}.json`)
-      .then((response) => {
-        const { data } = response;
-        let leadTime = 0;
-        const leadStatus = ['backlog', 'to do', 'open'];
-        const { histories } = data.change_log;
-        const IssueCreatedAt = data.created;
-        const currentStatus = data.status.name.toLowerCase();
-        const totalTime = moment().diff(moment(IssueCreatedAt), 'minutes');
-        const statusChanges = histories.filter(
-          (history) =>
-            ['status', 'resolution'].includes(history.items[0].field) &&
-            history.items[0].fieldtype === 'jira',
-        );
-
-        if (statusChanges.length === 0 && leadStatus.includes(currentStatus)) {
-          leadTime += totalTime;
-        }
-
-        statusChanges.forEach((statusChange, index) => {
-          const lastIndex = statusChange.items.length - 1;
-          const status = statusChange.items[lastIndex].toString
-            .toString()
-            .toLowerCase();
-
-          if (index === 0) {
-            const diff = moment().diff(moment(statusChange.created), 'minutes');
-
-            if (leadStatus.includes(status)) {
-              leadTime += diff;
-            }
-          } else if (index === statusChanges.length - 1) {
-            const diff = moment(statusChange.created).diff(
-              moment(IssueCreatedAt),
-              'minutes',
-            );
-
-            if (leadStatus.includes(status)) {
-              leadTime += diff;
-            }
-          } else if (index > 0 || index < statusChanges.length - 1) {
-            const createdAt = statusChanges[index + 1].created;
-            const diff = moment(statusChange.created).diff(
-              moment(createdAt),
-              'minutes',
-            );
-
-            if (leadStatus.includes(status)) {
-              leadTime += diff;
-            }
-          }
-        });
-
-        if (leadTime === 0 && leadStatus.includes(currentStatus)) {
-          leadTime += totalTime;
-        }
-
-        this.setState({
-          hoveredFocus: epicId,
-          leadTime: formattedTime(leadTime),
-        });
-      })
-      .finally(() => {});
-  };
-
-  renderFocus = (focus, index) => {
-    const colors = ['new-work', 'legacy-refactor', 'help-others', 'churn'];
-    const { epicIssues, leadTime, hoveredFocus } = this.state;
-    const filteredEpicIssues = epicIssues.filter(
-      (epicIssue) => epicIssue.epic_link === focus.key,
-    );
-    let percentage;
-    if (focus.status.name === 'Done') {
-      percentage = 100;
-    } else if (filteredEpicIssues.length > 0) {
-      const doneEpicIssues = filteredEpicIssues.filter(
-        (issue) => issue.status.name === 'Done',
-      );
-      const done = doneEpicIssues.length;
-      const total = filteredEpicIssues.length;
-      percentage = Math.round((done / total) * 100);
-    } else {
-      percentage = 0;
-    }
-
-    return (
-      <div
-        className="mb-4"
-        key={focus.key}
-        onClick={() => this.displayLeadTime(focus.issue_id)}
-        style={{ cursor: 'pointer' }}
-        aria-hidden="true"
-      >
-        <div className="mb-2 d-flex justify-content-between">
-          <div>{focus.epic_name || focus.key}</div>
-          <div className={`${hoveredFocus === focus.issue_id ? '' : 'd-none'}`}>
-            [Lead Time: {leadTime}]
-            <i
-              className="fa fa-times ml-1"
-              onClick={this.clearLeadTime}
-              aria-hidden="true"
-            />
-          </div>
-          <div className={`${hoveredFocus === focus.issue_id ? 'd-none' : ''}`}>
-            {percentage}%
-          </div>
-        </div>
-        <ProgressBar className={colors[index % 4]} now={percentage} />
-      </div>
-    );
-  };
-
-  renderAverageTimes = () => {
-    let leadTime = 0;
-    let developmentTime = 0;
-    let reviewTime = 0;
-    let deploymentTime = 0;
-    const { issues, selectedIssueId } = this.state;
-    const selectedIssue = issues.filter(
-      (issue) => issue.id === selectedIssueId,
-    );
-    const style = {
-      height: 300,
-      overflow: 'auto',
-    };
-    const leadStatus = ['backlog', 'to do', 'open'];
-    const devStatus = ['selected for development', 'in progress'];
-    const revStatus = ['review', 'qa', 'ready for review', 'in review'];
-    const deployStatus = ['done', 'closed'];
-
-    if (selectedIssue.length === 0) {
-      return <div />;
-    }
-
-    const { histories } = selectedIssue[0].change_log;
-    const IssueCreatedAt = selectedIssue[0].created;
-    const currentStatus = selectedIssue[0].status.name.toLowerCase();
-    const totalTime = moment().diff(
-      moment(selectedIssue[0].created),
-      'minutes',
-    );
-    const statusChanges = histories.filter(
-      (history) =>
-        ['status', 'resolution'].includes(history.items[0].field) &&
-        history.items[0].fieldtype === 'jira',
-    );
-
-    if (statusChanges.length === 0) {
-      if (leadStatus.includes(currentStatus)) {
-        leadTime += totalTime;
-      } else if (devStatus.includes(currentStatus)) {
-        developmentTime += totalTime;
-      } else if (revStatus.includes(currentStatus)) {
-        reviewTime += totalTime;
-      } else if (deployStatus.includes(currentStatus)) {
-        deploymentTime += totalTime;
-      }
-    }
-
-    statusChanges.forEach((statusChange, index) => {
-      const lastIndex = statusChange.items.length - 1;
-      const status = statusChange.items[lastIndex].toString
-        .toString()
-        .toLowerCase();
-
-      if (index === 0) {
-        const diff = moment().diff(moment(statusChange.created), 'minutes');
-
-        if (leadStatus.includes(status)) {
-          leadTime += diff;
-        } else if (devStatus.includes(status)) {
-          developmentTime += diff;
-        } else if (revStatus.includes(status)) {
-          reviewTime += diff;
-        } else if (deployStatus.includes(status)) {
-          deploymentTime += diff;
-        }
-      } else if (index === statusChanges.length - 1) {
-        const diff = moment(statusChange.created).diff(
-          moment(IssueCreatedAt),
-          'minutes',
-        );
-
-        if (leadStatus.includes(status)) {
-          leadTime += diff;
-        } else if (devStatus.includes(status)) {
-          developmentTime += diff;
-        } else if (revStatus.includes(status)) {
-          reviewTime += diff;
-        } else if (deployStatus.includes(status)) {
-          deploymentTime += diff;
-        }
-      } else if (index > 0 || index < statusChanges.length - 1) {
-        const createdAt = statusChanges[index + 1].created;
-        const diff = moment(statusChange.created).diff(
-          moment(createdAt),
-          'minutes',
-        );
-
-        if (leadStatus.includes(status)) {
-          leadTime += diff;
-        } else if (devStatus.includes(status)) {
-          developmentTime += diff;
-        } else if (revStatus.includes(status)) {
-          reviewTime += diff;
-        } else if (deployStatus.includes(status)) {
-          deploymentTime += diff;
-        }
-      }
-    });
-
-    if (
-      leadTime === 0 &&
-      developmentTime === 0 &&
-      reviewTime === 0 &&
-      deploymentTime === 0
-    ) {
-      if (leadStatus.includes(currentStatus)) {
-        leadTime += totalTime;
-      } else if (devStatus.includes(currentStatus)) {
-        developmentTime += totalTime;
-      } else if (revStatus.includes(currentStatus)) {
-        reviewTime += totalTime;
-      } else if (deployStatus.includes(currentStatus)) {
-        deploymentTime += totalTime;
-      }
-    }
-
-    const percentage = (time) => Math.floor((time / totalTime) * 100);
-
-    return (
-      <Card>
-        <Card.Body>
-          <Card.Title>Time Spent</Card.Title>
-          <div style={style}>
-            <div className="mb-4">
-              <div className="mb-2 d-flex justify-content-between avg-count">
-                <div>Lead Time</div>
-                <div className="le-ti">{formattedTime(leadTime)}</div>
-              </div>
-              <ProgressBar className="lead-time" now={percentage(leadTime)} />
-            </div>
-
-            <div className="mb-4">
-              <div className="mb-2 d-flex justify-content-between avg-count">
-                <div>Development Time</div>
-                <div className="dev">{formattedTime(developmentTime)}</div>
-              </div>
-              <ProgressBar
-                className="development"
-                now={percentage(developmentTime)}
-              />
-            </div>
-
-            <div className="mb-4">
-              <div className="mb-2 d-flex justify-content-between avg-count">
-                <div>Review Time</div>
-                <div className="rev">{formattedTime(reviewTime)}</div>
-              </div>
-              <ProgressBar className="review" now={percentage(reviewTime)} />
-            </div>
-
-            <div className="mb-4">
-              <div className="mb-2 d-flex justify-content-between avg-count">
-                <div>Deployment Time</div>
-                <div className="deploy">{formattedTime(deploymentTime)}</div>
-              </div>
-              <ProgressBar
-                className="deployment"
-                now={percentage(deploymentTime)}
-              />
-            </div>
-          </div>
-        </Card.Body>
-      </Card>
-    );
-  };
-
-  renderMinMaxIcon = (param) => {
-    if (param) return <i className="fa fa-minus" />;
-
-    return <i className="fa fa-plus" />;
-  };
-
-  minMaxHandler = (sectionName, value) => {
-    const {
-      showWIPSection,
-      showGaugeSection,
-      showFocusSection,
-      showVPISection,
-      showActivitiesSection,
-      showVSMSection,
-    } = this.state;
-    this.setState({ [sectionName]: !value });
-
-    const data = {
-      min_max: {
-        showWIPSection,
-        showGaugeSection,
-        showFocusSection,
-        showVPISection,
-        showActivitiesSection,
-        showVSMSection,
-        [sectionName]: !value,
-      },
-    };
-
-    axios.put('/users/min_max_config.json', data);
+    this.setState({ projectId: value });
   };
 
   onSortEnd = ({ oldIndex, newIndex }) => {
-    let { sortableItems } = this.state;
-
-    sortableItems = arrayMove(sortableItems, oldIndex, newIndex);
-
-    this.setState({ sortableItems });
-
-    axios.put('/users/sortable_items_config.json', {
-      sortable_items: sortableItems,
-    });
-  };
-
-  calculateVPI = (days, issues, rate) => {
-    if (days === null || issues === 0 || [0, null].includes(rate)) {
-      return '--';
+    if (oldIndex === newIndex) {
+      return;
     }
 
-    return (days / (issues * rate)).toFixed(2);
-  };
+    let { sortableItems } = this.state;
+    sortableItems = arrayMove(sortableItems, oldIndex, newIndex);
+    this.setState({ sortableItems });
 
-  projectName = (project_id) => {
-    const { projects } = this.state;
-    const filteredProject = projects.filter(
-      (project) => project.project_id === project_id,
-    );
-
-    return filteredProject[0].name;
+    const data = { user: { sortable_items: sortableItems } };
+    axios.put('/users/update.json', data);
   };
 
   render() {
     const {
-      issues,
-      epics,
-      bugs,
-      tasks,
       projects,
-      redirect,
-      devArcLength,
-      testArcLength,
-      deployArcLength,
-      jiraActivityLoading,
-      jiraActivityHasMore,
       showAlert,
       alertMessage,
-      selectedIssueId,
-      projectsSync,
-      issuesSync,
-      remaining_days,
-      remaining_issues,
-      average_time_to_close,
-      showVPI,
-      testVPI,
-      testAverageTimeToClose,
-      testRemainingIssues,
-      testRemainingDays,
-      totalBacklog,
-      totalWorkInProgress,
-      totalWorkInReview,
-      WIPChartType,
-      backlogLimit,
-      inProgressLimit,
-      inReviewLimit,
-      showWIPSection,
-      showGaugeSection,
-      showFocusSection,
-      showVPISection,
-      showActivitiesSection,
-      showVSMSection,
       sortableItems,
-      user,
-      showProjectVPIs,
-      showIngest,
-      projectVPIs,
-      statusLeadTimes,
-      statusProcessTimes,
-      statusCA,
-      VSMTotal,
-      showIngestAlert,
-      ingestErrorMessage,
-      disableIngest,
-      ingestType,
+      projectId,
+      initialConfig,
     } = this.state;
 
-    let listIssues;
-    let listEpics;
-    let listBugs;
-    let listTasks;
     let alert;
-    let VPI;
-    let healthRecommendation;
-    let WIPChart;
-    let VPIChart;
-    let VPIPercent;
-    let ingestAlert;
-    let ingestInfoMessage;
-
-    if (ingestType === 'issues') {
-      ingestInfoMessage =
-        'issue_id, project_id, status, issue_type, issue_key, change_log, summary, due_date & created';
-    } else if (ingestType === 'projects') {
-      ingestInfoMessage =
-        'board_id, name, board_type, project_id, column_config';
-    } else if (ingestType === 'boards') {
-      ingestInfoMessage = 'project_id, key & name';
-    }
-
-    if (jiraActivityLoading) {
-      listIssues = <Skeleton count={10} />;
-    } else if (issues.length === 0) {
-      listIssues = <div>No activities found.</div>;
-    } else {
-      listIssues = issues.map((issue) => (
-        <ListGroup.Item
-          key={issue.id}
-          className={`${
-            selectedIssueId === issue.id ? 'selected d-flex' : 'd-flex'
-          }`}
-          onClick={() => this.selectIssue(issue.id)}
-        >
-          <OverlayTrigger
-            placement="top"
-            overlay={
-              <Tooltip id={`tooltip-issue-type-${issue.id}`}>
-                {issue.issue_type.name}
-              </Tooltip>
-            }
-          >
-            <div>
-              <Image
-                src={issue.issue_type.icon_url}
-                alt={issue.issue_type.name}
-                rounded
-              />
-            </div>
-          </OverlayTrigger>
-          <div className="ml-2">
-            <strong>{issue.key}</strong>
-          </div>
-          <OverlayTrigger
-            placement="top"
-            overlay={
-              <Tooltip id={`tooltip-summary-${issue.id}`}>
-                {issue.summary}
-              </Tooltip>
-            }
-          >
-            <div className="text-truncate ml-2 issue-summary">
-              {issue.summary}
-            </div>
-          </OverlayTrigger>
-          <div className="ml-auto">
-            <Badge variant="secondary" className="text-uppercase">
-              {issue.status.name}
-            </Badge>
-          </div>
-        </ListGroup.Item>
-      ));
-    }
-
-    if (epics.length > 0) {
-      listEpics = epics.map((epic, index) => this.renderFocus(epic, index));
-    } else {
-      listEpics = <div>No epics found.</div>;
-    }
-
-    if (bugs.length > 0) {
-      listBugs = bugs.map((bug, index) => this.renderFocus(bug, index));
-    } else {
-      listBugs = <div>No bugs found.</div>;
-    }
-
-    if (tasks.length > 0) {
-      listTasks = tasks.map((task, index) => this.renderFocus(task, index));
-    } else {
-      listTasks = <div>No tasks found.</div>;
-    }
-
-    if (redirect) {
-      return <Redirect to="/login" />;
-    }
 
     if (showAlert) {
       const alertStyle = {
@@ -927,978 +96,17 @@ export default class Dashboard extends Component {
       );
     }
 
-    if (showIngestAlert) {
-      ingestAlert = (
-        <Alert variant="danger">
-          <span>{ingestErrorMessage}</span>
-        </Alert>
-      );
-    }
-
-    if (
-      remaining_days === null ||
-      remaining_issues === 0 ||
-      [0, null].includes(average_time_to_close)
-    ) {
-      VPI = '--';
-    } else {
-      VPI = (
-        remaining_days /
-        (remaining_issues * average_time_to_close)
-      ).toFixed(2);
-    }
-
-    if (remaining_days === null) {
-      healthRecommendation =
-        'Please assign due date at least for an issue to view the VPI score.';
-    } else if (remaining_issues === 0) {
-      healthRecommendation =
-        'There are no remaining issues to view the VPI score';
-    } else if (average_time_to_close === null) {
-      healthRecommendation =
-        'At least one issue needs to be completed to view the VPI score.';
-    } else if (average_time_to_close === 0) {
-      healthRecommendation =
-        'Average completion rate should be greater than zero to view the VPI score.';
-    } else if (VPI >= 0.67) {
-      healthRecommendation = 'VPI indicates early delivery.';
-      VPIPercent = 0.85;
-    } else if (VPI <= 0.33) {
-      healthRecommendation = 'VPI indicates late delivery.';
-      VPIPercent = 0.15;
-    } else {
-      healthRecommendation = 'VPI indicates on-schedule delivery.';
-      VPIPercent = 0.5;
-    }
-
-    if (WIPChartType === 'bar') {
-      const commonAnnotation = {
-        drawTime: 'afterDatasetsDraw',
-        type: 'box',
-        xScaleID: 'x-axis-0',
-        yScaleID: 'y-axis-0',
-        backgroundColor: 'rgba(101, 33, 171, 0.5)',
-        borderColor: 'rgb(101, 33, 171)',
-        borderWidth: 1,
-      };
-      const backlogAnnotation = {
-        ...commonAnnotation,
-        xMin: -0.4,
-        xMax: 0.4,
-        yMin: backlogLimit,
-        yMax: backlogLimit + 0.5,
-      };
-      const inProgressAnnotation = {
-        ...commonAnnotation,
-        xMin: 0.6,
-        xMax: 1.4,
-        yMin: inProgressLimit,
-        yMax: inProgressLimit + 0.5,
-      };
-      const inReviewAnnotation = {
-        ...commonAnnotation,
-        xMin: 1.6,
-        xMax: 2.4,
-        yMin: inReviewLimit,
-        yMax: inReviewLimit + 0.5,
-      };
-      WIPChart = (
-        <Bar
-          data={{
-            labels: [
-              `To Do ( WIP = ${totalBacklog} )`,
-              `In Progress ( WIP = ${totalWorkInProgress} )`,
-              `In Review ( WIP = ${totalWorkInReview} )`,
-            ],
-            datasets: [
-              {
-                data: [totalBacklog, totalWorkInProgress, totalWorkInReview],
-                backgroundColor: [
-                  'rgba(255, 99, 132, 0.2)',
-                  'rgba(54, 162, 235, 0.2)',
-                  'rgba(255, 206, 86, 0.2)',
-                ],
-                borderColor: [
-                  'rgba(255, 99, 132, 1)',
-                  'rgba(54, 162, 235, 1)',
-                  'rgba(255, 206, 86, 1)',
-                ],
-                borderWidth: 1,
-              },
-            ],
-          }}
-          width={100}
-          height={250}
-          plugins={[ChartAnnotation]}
-          options={{
-            annotation: {
-              annotations: [
-                backlogLimit === 0 ? {} : backlogAnnotation,
-                inProgressLimit === 0 ? {} : inProgressAnnotation,
-                inReviewLimit === 0 ? {} : inReviewAnnotation,
-              ],
-            },
-            maintainAspectRatio: false,
-            legend: {
-              display: false,
-            },
-            tooltips: {
-              enabled: true,
-              callbacks: {
-                label(tooltipItem, data) {
-                  const value = data.datasets[0].data[tooltipItem.index];
-                  let label = ` Work in progress: ${value}`;
-                  if (
-                    tooltipItem.label.toLowerCase() === 'to do' &&
-                    backlogLimit > 0
-                  ) {
-                    label += `; WIP Limit: ${backlogLimit}; Limit %: ${
-                      (value / backlogLimit).toFixed(2) * 100
-                    }`;
-                  } else if (
-                    tooltipItem.label.toLowerCase() === 'in progress' &&
-                    inProgressLimit > 0
-                  ) {
-                    label += `; WIP Limit: ${inProgressLimit}; Limit %: ${
-                      (value / inProgressLimit).toFixed(2) * 100
-                    }`;
-                  } else if (
-                    tooltipItem.label.toLowerCase() === 'in review' &&
-                    inReviewLimit > 0
-                  ) {
-                    label += `; WIP Limit: ${inReviewLimit}; Limit %: ${
-                      (value / inReviewLimit).toFixed(2) * 100
-                    }`;
-                  }
-
-                  return label;
-                },
-              },
-            },
-            scales: {
-              xAxes: [
-                {
-                  gridLines: {
-                    display: false,
-                  },
-                  ticks: {
-                    display: true,
-                  },
-                },
-              ],
-              yAxes: [
-                {
-                  gridLines: {
-                    display: false,
-                  },
-                  ticks: {
-                    display: false,
-                  },
-                },
-              ],
-            },
-          }}
-        />
-      );
-    } else if (WIPChartType === 'bubble') {
-      WIPChart = (
-        <Bubble
-          data={{
-            labels: ['To Do', 'In Progress', 'In Review'],
-            datasets: [
-              {
-                label: 'To Do',
-                data: [{ x: 0.5, y: 1, r: totalBacklog }],
-                backgroundColor: ['rgba(255, 99, 132, 0.2)'],
-                borderColor: ['rgba(255, 99, 132, 1)'],
-                borderWidth: 1,
-              },
-              {
-                label: 'To Do Limit',
-                data: [{ x: 0.5, y: 1, r: backlogLimit }],
-                backgroundColor: ['rgba(255, 255, 255, 0.2)'],
-                borderColor: ['rgba(255, 99, 132, 0.5)'],
-                borderWidth: 2,
-              },
-              {
-                label: 'In Progress',
-                data: [{ x: 1.5, y: 1, r: totalWorkInProgress }],
-                backgroundColor: ['rgba(54, 162, 235, 0.2)'],
-                borderColor: ['rgba(54, 162, 235, 1)'],
-                borderWidth: 1,
-              },
-              {
-                label: 'In Progress Limit',
-                data: [{ x: 1.5, y: 1, r: inProgressLimit }],
-                backgroundColor: ['rgba(255, 255, 255, 0.2)'],
-                borderColor: ['rgba(54, 162, 235, 0.5)'],
-                borderWidth: 2,
-              },
-              {
-                label: 'In Review',
-                data: [{ x: 2.5, y: 1, r: totalWorkInReview }],
-                backgroundColor: ['rgba(255, 206, 86, 0.2)'],
-                borderColor: ['rgba(255, 206, 86, 1)'],
-                borderWidth: 1,
-              },
-              {
-                label: 'In Review Limit',
-                data: [{ x: 2.5, y: 1, r: inReviewLimit }],
-                backgroundColor: ['rgba(255, 255, 255, 0.2)'],
-                borderColor: ['rgba(255, 206, 86, 0.5)'],
-                borderWidth: 2,
-              },
-            ],
-          }}
-          width={100}
-          height={250}
-          options={{
-            maintainAspectRatio: false,
-            legend: {
-              display: false,
-            },
-            tooltips: {
-              enabled: true,
-              callbacks: {
-                label(tooltipItem, data) {
-                  let label =
-                    data.datasets[tooltipItem.datasetIndex].label || '';
-
-                  if (label) {
-                    label += ': ';
-                  }
-                  label += data.datasets[tooltipItem.datasetIndex].data[0].r;
-                  return label;
-                },
-              },
-            },
-            scales: {
-              xAxes: [
-                {
-                  gridLines: {
-                    display: false,
-                  },
-                  ticks: {
-                    min: 0,
-                    display: false,
-                  },
-                },
-              ],
-              yAxes: [
-                {
-                  gridLines: {
-                    display: false,
-                  },
-                  ticks: {
-                    display: false,
-                  },
-                },
-              ],
-            },
-          }}
-        />
-      );
-    }
-
-    if (VPI !== '--') {
-      VPIChart = (
-        <GaugeChart
-          id="gauge_chart_vpi"
-          colors={['#ff736f', '#ffce00', '#49d1c5']}
-          cornerRadius={0}
-          arcWidth={0.1}
-          arcPadding={0.02}
-          percent={VPIPercent}
-          animate={false}
-          hideText
-          style={{ width: '75%' }}
-        />
-      );
-    }
-
-    const DragHandle = SortableHandle(() => (
-      <i className="fa fa-bars" style={{ cursor: 'move' }} />
-    ));
-
     const SortableItem = SortableElement(({ value }) => {
-      const sectionStyle = { position: 'relative' };
-      const helperStyle = {
-        position: 'absolute',
-        bottom: 0,
-        right: 0,
-        fontSize: '24px',
-        opacity: 0.5,
+      const sections = {
+        wip: <WIP projectId={projectId} />,
+        gauge: <Gauge projectId={projectId} />,
+        focus: <Focus projectId={projectId} />,
+        vpi: <VPI projectId={projectId} />,
+        activities: <Activities projectId={projectId} />,
+        vsm: <VSM projectId={projectId} />,
       };
 
-      if (value === 'wip') {
-        return (
-          <Row className="pt-4">
-            <Col xs={12}>
-              <Card>
-                <Card.Header
-                  className="d-flex justify-content-between align-items-center"
-                  onClick={() =>
-                    this.minMaxHandler('showWIPSection', showWIPSection)
-                  }
-                >
-                  <div className="d-inline-flex">
-                    <div className="mr-2">
-                      <DragHandle /> WIP + WIP Limits
-                    </div>
-                    <Form onClick={(event) => event.stopPropagation()}>
-                      <Form.Check
-                        inline
-                        label="Graph"
-                        type="radio"
-                        id="bar"
-                        name="WIP"
-                        onClick={this.WIPChartTypeChangeHandler}
-                        defaultChecked={WIPChartType === 'bar'}
-                      />
-                      <Form.Check
-                        inline
-                        label="Circle Chart"
-                        type="radio"
-                        id="bubble"
-                        name="WIP"
-                        onClick={this.WIPChartTypeChangeHandler}
-                        defaultChecked={WIPChartType === 'bubble'}
-                      />
-                    </Form>
-                  </div>
-                  {this.renderMinMaxIcon(showWIPSection)}
-                </Card.Header>
-                <Collapse in={showWIPSection} className="wip">
-                  <div style={sectionStyle}>
-                    <Card.Body>{WIPChart}</Card.Body>
-                    <OverlayTrigger
-                      key="wipHelper"
-                      placement="top"
-                      overlay={
-                        <Tooltip id="wipHelper">
-                          Work in Progress with limits
-                        </Tooltip>
-                      }
-                    >
-                      <i
-                        className="fa fa-question-circle m-3"
-                        style={helperStyle}
-                      />
-                    </OverlayTrigger>
-                  </div>
-                </Collapse>
-              </Card>
-            </Col>
-          </Row>
-        );
-      }
-
-      if (value === 'gauge') {
-        return (
-          <Row className="pt-4">
-            <Col xs={12}>
-              <Card>
-                <Card.Header
-                  className="d-flex justify-content-between align-items-center"
-                  onClick={() =>
-                    this.minMaxHandler('showGaugeSection', showGaugeSection)
-                  }
-                >
-                  <span>
-                    <DragHandle /> Status Gauges
-                  </span>
-                  {this.renderMinMaxIcon(showGaugeSection)}
-                </Card.Header>
-                <Collapse in={showGaugeSection}>
-                  <div style={sectionStyle}>
-                    <Card.Body>
-                      <Row>
-                        <Col xs={4}>
-                          <Card>
-                            <Card.Body>
-                              <Card.Title className="text-center">
-                                Backlog
-                              </Card.Title>
-                              <div className="d-flex align-items-center">
-                                <div className="legend left-legend" />
-                                <div>Backlog</div>
-                                <div className="ml-auto">Remaining</div>
-                                <div className="legend right-legend" />
-                              </div>
-                              <GaugeChart
-                                id="gauge_chart_dev"
-                                nrOfLevels={2}
-                                arcsLength={devArcLength}
-                                colors={['#e5e5e5', '#009cf0']}
-                                cornerRadius={0}
-                                arcWidth={0.1}
-                                arcPadding={0.02}
-                                hideText
-                                needleColor="#FFFFFF"
-                                needleBaseColor="#FFFFFF"
-                              />
-                            </Card.Body>
-                          </Card>
-                        </Col>
-                        <Col xs={4}>
-                          <Card>
-                            <Card.Body>
-                              <Card.Title className="text-center">
-                                Development/QA/Test
-                              </Card.Title>
-                              <div className="d-flex align-items-center">
-                                <div className="legend left-legend" />
-                                <div>In Progress</div>
-                                <div className="ml-auto">Remaining</div>
-                                <div className="legend right-legend" />
-                              </div>
-                              <GaugeChart
-                                id="gauge_chart_qa"
-                                nrOfLevels={2}
-                                arcsLength={testArcLength}
-                                colors={['#e5e5e5', '#009cf0']}
-                                cornerRadius={0}
-                                arcWidth={0.1}
-                                arcPadding={0.02}
-                                hideText
-                                needleColor="#FFFFFF"
-                                needleBaseColor="#FFFFFF"
-                              />
-                            </Card.Body>
-                          </Card>
-                        </Col>
-                        <Col xs={4}>
-                          <Card>
-                            <Card.Body>
-                              <Card.Title className="text-center">
-                                Deployment
-                              </Card.Title>
-                              <div className="d-flex align-items-center">
-                                <div className="legend left-legend" />
-                                <div>Done</div>
-                                <div className="ml-auto">Remaining</div>
-                                <div className="legend right-legend" />
-                              </div>
-                              <GaugeChart
-                                id="gauge_chart_deploy"
-                                nrOfLevels={2}
-                                arcsLength={deployArcLength}
-                                colors={['#e5e5e5', '#009cf0']}
-                                cornerRadius={0}
-                                arcWidth={0.1}
-                                arcPadding={0.02}
-                                hideText
-                                needleColor="#FFFFFF"
-                                needleBaseColor="#FFFFFF"
-                              />
-                            </Card.Body>
-                          </Card>
-                        </Col>
-                      </Row>
-                    </Card.Body>
-                    <OverlayTrigger
-                      key="gaugeHelper"
-                      placement="top"
-                      overlay={
-                        <Tooltip id="gaugeHelper">Status Gauges</Tooltip>
-                      }
-                    >
-                      <i
-                        className="fa fa-question-circle m-3"
-                        style={helperStyle}
-                      />
-                    </OverlayTrigger>
-                  </div>
-                </Collapse>
-              </Card>
-            </Col>
-          </Row>
-        );
-      }
-
-      if (value === 'focus') {
-        return (
-          <Row className="pt-4">
-            <Col xs={12}>
-              <Card>
-                <Card.Header
-                  className="d-flex justify-content-between align-items-center"
-                  onClick={() =>
-                    this.minMaxHandler('showFocusSection', showFocusSection)
-                  }
-                >
-                  <span>
-                    <DragHandle /> Focus
-                  </span>
-                  {this.renderMinMaxIcon(showFocusSection)}
-                </Card.Header>
-                <Collapse in={showFocusSection}>
-                  <div style={sectionStyle}>
-                    <Card.Body>
-                      <Row>
-                        <Col xs={4}>
-                          <Card>
-                            <Card.Body>
-                              <Card.Title>Epics</Card.Title>
-                              <div className="focus">{listEpics}</div>
-                            </Card.Body>
-                          </Card>
-                        </Col>
-                        <Col xs={4}>
-                          <Card>
-                            <Card.Body>
-                              <Card.Title>Bugs</Card.Title>
-                              <div className="focus">{listBugs}</div>
-                            </Card.Body>
-                          </Card>
-                        </Col>
-                        <Col xs={4}>
-                          <Card>
-                            <Card.Body>
-                              <Card.Title>Tasks</Card.Title>
-                              <div className="focus">{listTasks}</div>
-                            </Card.Body>
-                          </Card>
-                        </Col>
-                        {/* <Col xs={8}> */}
-                        {/*  <Card> */}
-                        {/*    <Card.Body> */}
-                        {/*      <Card.Title>Commit Breakdown</Card.Title> */}
-                        {/*      <div className="d-flex justify-content-around align-items-center"> */}
-                        {/*        <div className="commit-risk-chart-holder"> */}
-                        {/*          <Doughnut */}
-                        {/*            data={{ */}
-                        {/*              datasets: [ */}
-                        {/*                { */}
-                        {/*                  data: [60, 35, 5], */}
-                        {/*                  backgroundColor: [ */}
-                        {/*                    'rgb(94,196,182)', */}
-                        {/*                    'rgb(241,203,73)', */}
-                        {/*                    'rgb(254,139,169)', */}
-                        {/*                  ], */}
-                        {/*                }, */}
-                        {/*              ], */}
-                        {/*              labels: ['Low', 'Medium', 'High'], */}
-                        {/*            }} */}
-                        {/*            width={224} */}
-                        {/*            height={224} */}
-                        {/*            options={{ */}
-                        {/*              cutoutPercentage: 80, */}
-                        {/*              legend: { */}
-                        {/*                display: false, */}
-                        {/*              }, */}
-                        {/*              tooltips: { */}
-                        {/*                enabled: true, */}
-                        {/*              }, */}
-                        {/*            }} */}
-                        {/*          /> */}
-                        {/*          <div className="total-commit"> */}
-                        {/*            <div className="number">18</div> */}
-                        {/*            <div className="text">Total</div> */}
-                        {/*            <div className="text">Commits</div> */}
-                        {/*          </div> */}
-                        {/*        </div> */}
-                        {/*        <div className="risk-commit"> */}
-                        {/*          <div className="mb-4"> */}
-                        {/*            <div className="mb-2">Low Risk</div> */}
-                        {/*            <div className="d-flex align-items-center"> */}
-                        {/*              <div className="risk-commit-value low mr-2"> */}
-                        {/*                11 (61%) */}
-                        {/*              </div> */}
-                        {/*              <div className="mr-1 up text-success"> */}
-                        {/*                <i className="fa fa-arrow-up mr-1" /> */}
-                        {/*                83% */}
-                        {/*              </div> */}
-                        {/*              <div className="text-secondary"> */}
-                        {/*                since last PI */}
-                        {/*              </div> */}
-                        {/*            </div> */}
-                        {/*          </div> */}
-                        {/*          <div className="mb-4"> */}
-                        {/*            <div className="mb-2">Medium Risk</div> */}
-                        {/*            <div className="d-flex align-items-center"> */}
-                        {/*              <div className="risk-commit-value medium mr-2"> */}
-                        {/*                7 (39%) */}
-                        {/*              </div> */}
-                        {/*              <div className="mr-1 down text-danger"> */}
-                        {/*                <i className="fa fa-arrow-down mr-1" /> */}
-                        {/*                82% */}
-                        {/*              </div> */}
-                        {/*              <div className="text-secondary"> */}
-                        {/*                since last PI */}
-                        {/*              </div> */}
-                        {/*            </div> */}
-                        {/*          </div> */}
-                        {/*          <div> */}
-                        {/*            <div className="mb-2">High Risk</div> */}
-                        {/*            <div className="d-flex align-items-center"> */}
-                        {/*              <div className="risk-commit-value high mr-2"> */}
-                        {/*                1 (0%) */}
-                        {/*              </div> */}
-                        {/*              <div className="mr-1 down text-danger"> */}
-                        {/*                <i className="fa fa-arrow-down mr-1" /> */}
-                        {/*                100% */}
-                        {/*              </div> */}
-                        {/*              <div className="text-secondary"> */}
-                        {/*                since last PI */}
-                        {/*              </div> */}
-                        {/*            </div> */}
-                        {/*          </div> */}
-                        {/*        </div> */}
-                        {/*      </div> */}
-                        {/*    </Card.Body> */}
-                        {/*  </Card> */}
-                        {/* </Col> */}
-                      </Row>
-                    </Card.Body>
-                    <OverlayTrigger
-                      key="focusHelper"
-                      placement="top"
-                      overlay={
-                        <Tooltip id="focusHelper">
-                          Focus - Epics, Bugs & Tasks
-                        </Tooltip>
-                      }
-                    >
-                      <i
-                        className="fa fa-question-circle m-3"
-                        style={helperStyle}
-                      />
-                    </OverlayTrigger>
-                  </div>
-                </Collapse>
-              </Card>
-            </Col>
-          </Row>
-        );
-      }
-
-      if (value === 'vpi') {
-        return (
-          <Row className="pt-4">
-            <Col xs={12}>
-              <Card>
-                <Card.Header
-                  className="d-flex justify-content-between align-items-center"
-                  onClick={() =>
-                    this.minMaxHandler('showVPISection', showVPISection)
-                  }
-                >
-                  <span>
-                    <DragHandle /> Project Flow Health
-                  </span>
-                  {this.renderMinMaxIcon(showVPISection)}
-                </Card.Header>
-                <Collapse in={showVPISection}>
-                  <div style={sectionStyle}>
-                    <Card.Body className="text-center">
-                      <Card.Text className="text-info">
-                        {healthRecommendation}
-                      </Card.Text>
-                      <div className="d-flex align-items-center justify-content-around">
-                        <div className="pt-2">
-                          <div style={{ fontSize: '32px' }}>
-                            {remaining_days === null ? '--' : remaining_days}
-                          </div>
-                          <div className="mt-1">Remaining Period (Days)</div>
-                        </div>
-                        <div className="pt-2">
-                          <div style={{ fontSize: '32px' }}>
-                            {remaining_issues}
-                          </div>
-                          <div className="mt-1">Remaining Issues</div>
-                        </div>
-                        <div className="pt-2">
-                          <div style={{ fontSize: '32px' }}>
-                            {average_time_to_close === null
-                              ? '--'
-                              : average_time_to_close}
-                          </div>
-                          <div className="mt-1">
-                            Average Completion Rate (Days)
-                          </div>
-                        </div>
-                        <div className="pt-2">
-                          {VPIChart}
-                          <div style={{ fontSize: '32px' }}>{VPI}</div>
-                          <div className="mt-1">VPI</div>
-                        </div>
-                      </div>
-                    </Card.Body>
-                    <OverlayTrigger
-                      key="vpiHelper"
-                      placement="top"
-                      overlay={
-                        <Tooltip id="vpiHelper">Project Flow Health</Tooltip>
-                      }
-                    >
-                      <i
-                        className="fa fa-question-circle m-3"
-                        style={helperStyle}
-                      />
-                    </OverlayTrigger>
-                  </div>
-                </Collapse>
-              </Card>
-            </Col>
-          </Row>
-        );
-      }
-
-      if (value === 'activities') {
-        const activitiesStyle = {
-          height: 300,
-          overflow: 'auto',
-        };
-
-        return (
-          <Row className="py-4">
-            <Col xs={12}>
-              <Card>
-                <Card.Header
-                  className="d-flex justify-content-between align-items-center"
-                  onClick={() =>
-                    this.minMaxHandler(
-                      'showActivitiesSection',
-                      showActivitiesSection,
-                    )
-                  }
-                >
-                  <span>
-                    <DragHandle /> Activities & Time Spent
-                  </span>
-                  {this.renderMinMaxIcon(showActivitiesSection)}
-                </Card.Header>
-                <Collapse in={showActivitiesSection}>
-                  <div style={sectionStyle}>
-                    <Card.Body>
-                      <Row>
-                        <Col xs={8}>
-                          <Card>
-                            <Card.Body>
-                              <Card.Title>Activities</Card.Title>
-                              <ListGroup
-                                variant="flush"
-                                style={activitiesStyle}
-                                id="jiraActivity"
-                              >
-                                <InfiniteScroll
-                                  dataLength={issues.length}
-                                  next={this.fetchMoreData}
-                                  hasMore={jiraActivityHasMore}
-                                  loader={<Skeleton count={10} />}
-                                  scrollableTarget="jiraActivity"
-                                >
-                                  {listIssues}
-                                </InfiniteScroll>
-                              </ListGroup>
-                            </Card.Body>
-                          </Card>
-                        </Col>
-                        <Col xs={4}>{this.renderAverageTimes()}</Col>
-                      </Row>
-                    </Card.Body>
-                    <OverlayTrigger
-                      key="activitiesHelper"
-                      placement="top"
-                      overlay={
-                        <Tooltip id="activitiesHelper">
-                          Activities & Time Spent
-                        </Tooltip>
-                      }
-                    >
-                      <i
-                        className="fa fa-question-circle m-3"
-                        style={helperStyle}
-                      />
-                    </OverlayTrigger>
-                  </div>
-                </Collapse>
-              </Card>
-            </Col>
-          </Row>
-        );
-      }
-
-      if (value === 'vsm') {
-        const statusStyle = {
-          padding: '10px',
-          border: 'solid 3px rgba(0, 0, 0, 0.4)',
-          borderRadius: '5px',
-          minWidth: '160px',
-          textAlign: 'center',
-          backgroundColor: '#e3e6e9',
-        };
-        const nonStatusStyle = {
-          ...statusStyle,
-          backgroundColor: '#fff',
-        };
-        const statHolderStyle = {
-          width: '130px',
-          margin: '0 15px',
-        };
-        const statStyle = {
-          border: 'solid 1px #ccc',
-          borderTop: 'none',
-          borderRadius: '5px',
-          padding: '8px',
-        };
-        const VSMArrow = {
-          fontSize: '42px',
-          color: '#9c9ea0',
-        };
-        const totalStyle = {
-          border: 'solid 1px #ccc',
-          borderRadius: '5px',
-          padding: '8px',
-        };
-        const startEndStyle = {
-          margin: '0 22px',
-        };
-
-        return (
-          <Row className="py-4">
-            <Col xs={12}>
-              <Card>
-                <Card.Header
-                  className="d-flex justify-content-between align-items-center"
-                  onClick={() =>
-                    this.minMaxHandler('showVSMSection', showVSMSection)
-                  }
-                >
-                  <span>
-                    <DragHandle /> VSM
-                  </span>
-                  {this.renderMinMaxIcon(showVSMSection)}
-                </Card.Header>
-                <Collapse in={showVSMSection}>
-                  <div style={sectionStyle}>
-                    <Card.Body>
-                      <div
-                        className="d-flex justify-content-between"
-                        style={startEndStyle}
-                      >
-                        <div className="text-center">
-                          <div style={nonStatusStyle}>Request</div>
-                          <i
-                            className="fa fa-long-arrow-down m-3"
-                            style={VSMArrow}
-                          />
-                        </div>
-                        <div className="text-center">
-                          <div style={nonStatusStyle}>User</div>
-                          <i
-                            className="fa fa-long-arrow-up m-3"
-                            style={VSMArrow}
-                          />
-                        </div>
-                      </div>
-                      <div className="d-flex justify-content-around">
-                        <div>
-                          <div style={statusStyle}>To Do</div>
-                          <div style={statHolderStyle}>
-                            <div className="text-center" style={statStyle}>
-                              LT = {statusLeadTimes.to_do} days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              PT = 0 days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              %C/A = {statusCA.to_do} %
-                            </div>
-                          </div>
-                        </div>
-                        <i
-                          className="fa fa-long-arrow-right"
-                          style={VSMArrow}
-                        />
-                        <div>
-                          <div style={statusStyle}>In Progress</div>
-                          <div style={statHolderStyle}>
-                            <div className="text-center" style={statStyle}>
-                              LT = {statusLeadTimes.wip} days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              PT = {statusProcessTimes.wip} days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              %C/A = {statusCA.wip} %
-                            </div>
-                          </div>
-                        </div>
-                        <i
-                          className="fa fa-long-arrow-right"
-                          style={VSMArrow}
-                        />
-                        <div>
-                          <div style={statusStyle}>In Review</div>
-                          <div style={statHolderStyle}>
-                            <div className="text-center" style={statStyle}>
-                              LT = {statusLeadTimes.qa} days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              PT = {statusProcessTimes.qa} days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              %C/A = {statusCA.qa} %
-                            </div>
-                          </div>
-                        </div>
-                        <i
-                          className="fa fa-long-arrow-right"
-                          style={VSMArrow}
-                        />
-                        <div>
-                          <div style={statusStyle}>Done</div>
-                          <div style={statHolderStyle}>
-                            <div className="text-center" style={statStyle}>
-                              LT = 0 days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              PT = 0 days
-                            </div>
-                            <div className="text-center" style={statStyle}>
-                              %C/A = 0 %
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                      <div className="d-flex justify-content-around mt-3">
-                        <div style={totalStyle}>
-                          Total LT = {VSMTotal.total_lt} days
-                        </div>
-                        <div style={totalStyle}>
-                          Total PT = {VSMTotal.total_pt} days
-                        </div>
-                        <div style={totalStyle}>
-                          Activity Ratio = {VSMTotal.activity_ratio} %
-                        </div>
-                        <div style={totalStyle}>
-                          Rolled %C/A = {VSMTotal.rolled_ca} %
-                        </div>
-                      </div>
-                    </Card.Body>
-                    <OverlayTrigger
-                      key="vsmHelper"
-                      placement="top"
-                      overlay={
-                        <Tooltip id="vsmHelper">Value Stream Map</Tooltip>
-                      }
-                    >
-                      <i
-                        className="fa fa-question-circle m-3"
-                        style={helperStyle}
-                      />
-                    </OverlayTrigger>
-                  </div>
-                </Collapse>
-              </Card>
-            </Col>
-          </Row>
-        );
-      }
-
-      return '';
+      return sections[value];
     });
 
     const SortableList = SortableContainer(({ items }) => {
@@ -1911,234 +119,18 @@ export default class Dashboard extends Component {
       );
     });
 
+    if (!initialConfig) {
+      return <Redirect to="/initial-config-step-1" />;
+    }
+
+    if (projects.length <= 0) {
+      return 'Loading...';
+    }
+
     return (
-      <section>
+      <section className="pb-4">
         {alert}
-
-        <nav>
-          <span>
-            <Figure className="m-0">
-              <Figure.Image
-                src="https://user-images.githubusercontent.com/38546045/87486176-f1a5f280-c5f7-11ea-90de-1e80393d15a0.png"
-                width={50}
-                height={30}
-                alt="Logo"
-                className="m-0"
-              />
-            </Figure>
-          </span>
-
-          <span>
-            <Dropdown>
-              <Dropdown.Toggle variant="primary" id="navDropdown">
-                <i className="fa fa-bars" />
-              </Dropdown.Toggle>
-
-              <Dropdown.Menu>
-                <Figure className="m-0 d-flex justify-content-center">
-                  <Figure.Image
-                    src="/avatar.jpg"
-                    width={50}
-                    height={30}
-                    alt="Avatar"
-                    className="m-0"
-                  />
-                </Figure>
-                <Figure.Caption className="text-center">
-                  {user.email}
-                </Figure.Caption>
-                <Dropdown.Divider />
-                <Dropdown.Item href="#" onClick={this.handleIngestShow}>
-                  Ingest .csv
-                </Dropdown.Item>
-                <Dropdown.Item href="#" onClick={this.handleVPIShow}>
-                  VPI Calculator
-                </Dropdown.Item>
-                <Dropdown.Item href="#" onClick={this.handleProjectVPIsShow}>
-                  VPI By Project
-                </Dropdown.Item>
-                <Dropdown.Item
-                  href="#"
-                  onClick={this.syncProjects}
-                  disabled={projectsSync}
-                >
-                  Sync Projects{' '}
-                  {projectsSync ? (
-                    <Spinner
-                      as="span"
-                      animation="border"
-                      size="sm"
-                      role="status"
-                      aria-hidden="true"
-                    />
-                  ) : (
-                    ''
-                  )}
-                </Dropdown.Item>
-                <Dropdown.Item
-                  href="#"
-                  onClick={this.syncIssues}
-                  disabled={issuesSync}
-                >
-                  Sync Issues{' '}
-                  {issuesSync ? (
-                    <Spinner
-                      as="span"
-                      animation="border"
-                      size="sm"
-                      role="status"
-                      aria-hidden="true"
-                    />
-                  ) : (
-                    ''
-                  )}
-                </Dropdown.Item>
-                <NavLink to="/profile">
-                  <Dropdown.Item as="span">Profile</Dropdown.Item>
-                </NavLink>
-                <Dropdown.Item href="#" onClick={this.logoutUser}>
-                  Logout
-                </Dropdown.Item>
-              </Dropdown.Menu>
-            </Dropdown>
-          </span>
-        </nav>
-
-        <Modal
-          show={showVPI}
-          onHide={this.handleVPIClose}
-          backdrop="static"
-          keyboard={false}
-        >
-          <Modal.Header closeButton>
-            <Modal.Title>VPI Test Harness</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <Form.Group>
-              <Form.Label>Remaining Period (Days)</Form.Label>
-              <Form.Control
-                type="number"
-                value={testRemainingDays}
-                onChange={this.handleRD}
-              />
-            </Form.Group>
-            <Form.Group>
-              <Form.Label>Remaining Issues</Form.Label>
-              <Form.Control
-                type="number"
-                value={testRemainingIssues}
-                onChange={this.handleRI}
-              />
-            </Form.Group>
-            <Form.Group>
-              <Form.Label>Average Completion Rate (Days)</Form.Label>
-              <Form.Control
-                type="number"
-                value={testAverageTimeToClose}
-                onChange={this.handleACR}
-              />
-            </Form.Group>
-            <Form.Group>
-              <Form.Label>Value Performance Index</Form.Label>
-              <div className="text-center" style={{ fontSize: '32px' }}>
-                {testVPI}
-              </div>
-            </Form.Group>
-          </Modal.Body>
-        </Modal>
-
-        <Modal
-          show={showProjectVPIs}
-          onHide={this.handleProjectVPIsClose}
-          backdrop="static"
-          keyboard={false}
-          size="lg"
-        >
-          <Modal.Header closeButton>
-            <Modal.Title>VPI By Project</Modal.Title>
-          </Modal.Header>
-          <Modal.Body style={{ maxHeight: '500px', overflowY: 'scroll' }}>
-            <Table striped bordered hover>
-              <thead>
-                <tr>
-                  <th>Project</th>
-                  <th>Remaining Period (Days)</th>
-                  <th>Remaining Issues</th>
-                  <th>Average Completion Rate (Days)</th>
-                  <th>VPI</th>
-                </tr>
-              </thead>
-              <tbody>
-                {projectVPIs.map((vpi) => (
-                  <tr key={vpi.project_id}>
-                    <td>{this.projectName(vpi.project_id)}</td>
-                    <td>
-                      {vpi.remaining_days === null ? '--' : vpi.remaining_days}
-                    </td>
-                    <td>{vpi.remaining_issues}</td>
-                    <td>
-                      {vpi.average_time_to_close === null
-                        ? '--'
-                        : vpi.average_time_to_close}
-                    </td>
-                    <td>
-                      {this.calculateVPI(
-                        vpi.remaining_days,
-                        vpi.remaining_issues,
-                        vpi.average_time_to_close,
-                      )}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </Table>
-          </Modal.Body>
-        </Modal>
-
-        <Modal
-          show={showIngest}
-          onHide={this.handleIngestClose}
-          backdrop="static"
-          keyboard={false}
-          size="md"
-        >
-          <Modal.Header closeButton>
-            <Modal.Title>Ingest .csv</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <Form>
-              {ingestAlert}
-              <Form.Control
-                as="select"
-                size="md"
-                onChange={this.handleIngestTypeChange}
-              >
-                <option value="projects">Projects</option>
-                <option value="issues">Issues</option>
-                <option value="boards">Boards</option>
-              </Form.Control>
-              <Form.File
-                name="ingest"
-                onChange={this.handleIngestChange}
-                accept=".csv"
-                className="mt-3"
-              />
-              <Button
-                variant="outline-primary"
-                type="submit"
-                onClick={this.handleSubmit}
-                disabled={disableIngest}
-                className="mt-3"
-              >
-                Ingest
-              </Button>
-              <Alert variant="info" className="mt-3">
-                For {ingestType}, we expect headers - {ingestInfoMessage}.
-              </Alert>
-            </Form>
-          </Modal.Body>
-        </Modal>
-
+        <Navbar projects={projects} />
         <Container className="pt-5 mb-5">
           <Row className="pt-4">
             <Col xs={12}>
@@ -2163,55 +155,7 @@ export default class Dashboard extends Component {
             onSortEnd={this.onSortEnd}
             useDragHandle
           />
-          {/* <Row className="pt-4"> */}
-          {/*  <Col xs={3}> */}
-          {/*    <Card> */}
-          {/*      <Card.Body> */}
-          {/*        <ProgressBar className="lead-time" now={100} /> */}
-          {/*        <div className="mb-2">Lead Time</div> */}
-          {/*        <div className="avg-count"> */}
-          {/*          <span className="le-ti">14 days</span> avg. */}
-          {/*        </div> */}
-          {/*      </Card.Body> */}
-          {/*    </Card> */}
-          {/*  </Col> */}
-          {/*  <Col xs={3}> */}
-          {/*    <Card> */}
-          {/*      <Card.Body> */}
-          {/*        <ProgressBar className="development" now={100} /> */}
-          {/*        <div className="mb-2">Development</div> */}
-          {/*        <div className="avg-count"> */}
-          {/*          <span className="dev">2.3 day</span> avg. */}
-          {/*        </div> */}
-          {/*      </Card.Body> */}
-          {/*    </Card> */}
-          {/*  </Col> */}
-
-          {/*  <Col xs={3}> */}
-          {/*    <Card> */}
-          {/*      <Card.Body> */}
-          {/*        <ProgressBar className="review" now={100} /> */}
-          {/*        <div className="mb-2">Review</div> */}
-          {/*        <div className="avg-count"> */}
-          {/*          <span className="rev">12.4 hours</span> avg. */}
-          {/*        </div> */}
-          {/*      </Card.Body> */}
-          {/*    </Card> */}
-          {/*  </Col> */}
-          {/*  <Col xs={3}> */}
-          {/*    <Card> */}
-          {/*      <Card.Body> */}
-          {/*        <ProgressBar className="deployment" now={100} /> */}
-          {/*        <div className="mb-2">Deployment</div> */}
-          {/*        <div className="avg-count"> */}
-          {/*          <span className="deploy">2.7 week</span> avg. */}
-          {/*        </div> */}
-          {/*      </Card.Body> */}
-          {/*    </Card> */}
-          {/*  </Col> */}
-          {/* </Row> */}
         </Container>
-
         <Footer />
       </section>
     );
