+++ front/js/components/Nav/Nav.js
@@ -11,7 +11,7 @@ const Nav = (props) => {
             <img
               alt="logo"
               className="logo-nav"
-              src="https://user-images.githubusercontent.com/38546045/87486176-f1a5f280-c5f7-11ea-90de-1e80393d15a0.png"
+              src="/logo.png"
             />
             <span className="top-version">Alpha</span>
           </Link>
