+++ front/js/components/Register/Register.js
@@ -114,7 +114,7 @@ export default class Register extends Component {
               <div className="mb-3 d-flex justify-content-center">
                 <img
                   alt="register"
-                  src="https://user-images.githubusercontent.com/38546045/87486176-f1a5f280-c5f7-11ea-90de-1e80393d15a0.png"
+                  src="/logo.png"
                   width="75px"
                   height="75px"
                 />
