+++ front/js/components/Settings/IngestProjects.js
@@ -99,7 +99,7 @@ export default class IngestProjects extends Component {
           <span>
             <Figure className="m-0">
               <Figure.Image
-                src="https://user-images.githubusercontent.com/38546045/87486176-f1a5f280-c5f7-11ea-90de-1e80393d15a0.png"
+                src="/logo.png"
                 width={50}
                 height={30}
                 alt="Logo"
